+++ src/vanishcat/config.py
@@ -1,0 +1,47 @@
+# -*- coding: utf-8 -*-
+"""Config: vanish cat
+"""
+import os
+import sys
+sys.path.append(os.path.join(os.path.dirname(__file__), '../..'))
+sys.path.append('storybuilder')
+
+
+# configs
+CHARAS = (
+        # main
+        # sub
+        # mob
+        )
+
+STAGES = (
+        # Area
+        # Place
+        # Part
+        )
+
+DAYS = (
+        # main
+        ("current", "現在"),
+        # sub
+        )
+
+ITEMS = (
+        # main
+        # sub
+        )
+
+INFOS = (
+        # theme
+        # main
+        # sub
+        )
+
+FLAGS = (
+        )
+
+THEMES = {
+        }
+
+MOTIFS = {
+        }